more declaration cleanup
