+++ fq_nmod_mpoly_factor/n_bpoly_hlift.c
@@ -20,483 +20,6 @@
         1: successfully lifted B0 and B1 to true factors without changing lc_x
 */
 
-#define MAC(h, m, l, a, b)                          \
-{                                                   \
-    mp_limb_t p1, p0;                               \
-    umul_ppmm(p1, p0, a, b);                        \
-    add_sssaaaaaa(h, m, l, h, m, l, 0, p1, p0);     \
-}
-
-
-static void _from_coeffs2_n_fq(
-    mp_limb_t * v,          /* length 4D+1 */
-    const mp_limb_t * a,    /* length alen <= 2D+1 */
-    slong alen,
-    const mp_limb_t * M_,    /* length D by 4D */
-    slong D,
-    slong d,
-    nmod_t ctx)
-{
-    slong i, j, k;
-    const mp_limb_t * Mrow;
-
-    FLINT_ASSERT(0 <= alen);
-    FLINT_ASSERT(alen <= 1 + 2*D);
-
-    if (alen <= 1)
-    {
-        if (alen == 1)
-        {
-            for (i = 0; i < 2*D+1; i++)
-                _n_fq_set(v + d*i, a + d*0, d);
-        }
-        else
-        {
-            _nmod_vec_zero(v, d*(2*D+1));
-        }
-
-        return;
-    }
-
-    _n_fq_set(v + d*0, a + d*0, d);
-
-    for (k = 0; k < d; k++)
-    {
-        Mrow = M_;
-        for (i = 0; i < D; i++)
-        {
-            mp_limb_t c1h, c1m, c1;
-            mp_limb_t c2h, c2m, c2;
-            c1h = c1m = c1 = 0;
-            c2h = c2m = c2 = 0;
-
-            for (j = 0; j + 2 < alen; j += 2)
-            {
-                FLINT_ASSERT(j + 1 < 4*D);
-                MAC(c1h, c1m, c1, a[d*(j+1)+k], Mrow[j+0]);
-                MAC(c2h, c2m, c2, a[d*(j+2)+k], Mrow[j+1]);
-            }
-
-            MAC(c1h, c1m, c1, j + 1 < alen ? a[d*(j+1)+k] : 0, Mrow[j+0]);
-            FLINT_ASSERT(j + 2 >= alen);
-
-            NMOD_RED3(c2, c2h, c2m, c2, ctx);
-            NMOD_RED3(c1, c1h, c1m, c1, ctx);
-
-            Mrow += 2*D;
-
-            c2 = nmod_add(c2, a[d*0+k], ctx);
-
-            v[d*(2*i+1)+k] = nmod_add(c2, c1, ctx);
-            v[d*(2*i+2)+k] = nmod_sub(c2, c1, ctx);
-        }
-    }
-}
-
-static void _to_coeffs2_n_fq(
-    mp_limb_t * a,          /* length 2d+1 */
-    const mp_limb_t * v,    /* length 2d+1 */
-    mp_limb_t * t,          /* length 2d   */
-    const mp_limb_t * Q_,    /* length d by 2d+1 */
-    slong D,
-    slong d,
-    nmod_t ctx)
-{
-    slong i, j, k;
-    const mp_limb_t * Qrow;
-
-    _n_fq_set(a + d*0, v + d*0, d);
-
-    for (k = 0; k < d; k++)
-    {
-        for (i = 0; i < D; i++)
-        {
-            t[2*i+0] = nmod_sub(v[d*(1+2*i+0)+k], v[d*(1+2*i+1)+k], ctx);
-            t[2*i+1] = nmod_add(v[d*(1+2*i+0)+k], v[d*(1+2*i+1)+k], ctx);
-        }
-
-        Qrow = Q_;
-        for (i = 0; i < D; i++)
-        {
-            mp_limb_t c1h, c1m, c1;
-            mp_limb_t c2h, c2m, c2;
-            c1h = c1m = c1 = 0;
-            c2h = c2m = c2 = 0;
-            umul_ppmm(c2m, c2, Qrow[0], v[d*0+k]);
-            for (j = 0; j < D; j++)
-            {
-                MAC(c1h, c1m, c1, t[2*j+0], Qrow[2*j+1]);
-                MAC(c2h, c2m, c2, t[2*j+1], Qrow[2*j+2]);
-            }
-
-            Qrow += 2*D + 1;
-
-            NMOD_RED3(a[d*(2*i+1)+k], c1h, c1m, c1, ctx);
-            NMOD_RED3(a[d*(2*i+2)+k], c2h, c2m, c2, ctx);
-        }
-    }
-}
-
-
-static void _from_coeffs4_n_fq(
-    mp_limb_t * v,          /* length 4d+1 */
-    const mp_limb_t * a,
-    slong alen,
-    const mp_limb_t * M_,    /* length d by 4d */
-    slong D,
-    mp_limb_t w,
-    slong d,
-    nmod_t ctx)
-{
-    slong i, j, k;
-    const mp_limb_t * Mrow;
-
-    FLINT_ASSERT(0 <= alen);
-    FLINT_ASSERT(alen <= 1 + 4*D);
-
-    if (alen <= 1)
-    {
-        if (alen == 1)
-        {
-            for (i = 0; i < 4*D + 1; i++)
-                _n_fq_set(v + d*i, a + d*0, d);
-        }
-        else
-        {
-            _nmod_vec_zero(v, d*(4*D+1));
-        }
-
-        return;
-    }
-
-    _n_fq_set(v + d*0, a + d*0, d);
-
-    for (k = 0; k < d; k++)
-    {
-        Mrow = M_;
-        for (i = 0; i < D; i++)
-        {
-            mp_limb_t t1, t2, t3, t4;
-            mp_limb_t c1h, c1m, c1;
-            mp_limb_t c2h, c2m, c2;
-            mp_limb_t c3h, c3m, c3;
-            mp_limb_t c4h, c4m, c4;
-            c1h = c1m = c1 = 0;
-            c2h = c2m = c2 = 0;
-            c3h = c3m = c3 = 0;
-            c4h = c4m = c4 = 0;
-
-            for (j = 0; j + 4 < alen; j += 4)
-            {
-                FLINT_ASSERT(j + 3 < 4*D);
-                MAC(c1h, c1m, c1, a[d*(j+1)+k], Mrow[j+0]);
-                MAC(c2h, c2m, c2, a[d*(j+2)+k], Mrow[j+1]);
-                MAC(c3h, c3m, c3, a[d*(j+3)+k], Mrow[j+2]);
-                MAC(c4h, c4m, c4, a[d*(j+4)+k], Mrow[j+3]);
-            }
-
-            MAC(c1h, c1m, c1, j + 1 < alen ? a[d*(j+1)+k] : 0, Mrow[j+0]);
-            MAC(c2h, c2m, c2, j + 2 < alen ? a[d*(j+2)+k] : 0, Mrow[j+1]);
-            MAC(c3h, c3m, c3, j + 3 < alen ? a[d*(j+3)+k] : 0, Mrow[j+2]);
-            FLINT_ASSERT(j + 4 >= alen);
-
-            NMOD_RED3(c4, c4h, c4m, c4, ctx);
-            NMOD_RED3(c1, c1h, c1m, c1, ctx);
-            NMOD_RED3(c2, c2h, c2m, c2, ctx);
-            NMOD_RED3(c3, c3h, c3m, c3, ctx);
-
-            Mrow += 4*D;
-
-            c4 = nmod_add(c4, a[d*0+k], ctx);
-            t1 = nmod_add(c4, c2, ctx);
-            t2 = nmod_sub(c4, c2, ctx);
-            t3 = nmod_add(c1, c3, ctx);
-            t4 = nmod_mul(nmod_sub(c1, c3, ctx), w, ctx);
-
-            v[d*(4*i+1)+k] = nmod_add(t1, t3, ctx);
-            v[d*(4*i+2)+k] = nmod_add(t2, t4, ctx);
-            v[d*(4*i+3)+k] = nmod_sub(t1, t3, ctx);
-            v[d*(4*i+4)+k] = nmod_sub(t2, t4, ctx);
-        }
-    }
-}
-
-static void _to_coeffs4_n_fq(
-    mp_limb_t * a,          /* length 4D+1 */
-    const mp_limb_t * v,    /* length 4D+1 */
-    mp_limb_t * t,          /* length 4D   */
-    const mp_limb_t * Q_,   /* length D by 4D+1 */
-    slong D,
-    mp_limb_t w,
-    slong d,
-    nmod_t ctx)
-{
-    slong i, j, k;
-    const mp_limb_t * Qrow;
-
-    _n_fq_set(a + d*0, v + d*0, d);
-
-    for (k = 0; k < d; k++)
-    {
-
-        for (i = 0; i < D; i++)
-        {
-            mp_limb_t t2 = nmod_add(v[d*(1+4*i+0)+k], v[d*(1+4*i+2)+k], ctx);
-            mp_limb_t t1 = nmod_sub(v[d*(1+4*i+0)+k], v[d*(1+4*i+2)+k], ctx);
-            mp_limb_t t3 = nmod_add(v[d*(1+4*i+1)+k], v[d*(1+4*i+3)+k], ctx);
-            mp_limb_t t4 = nmod_mul(nmod_sub(v[d*(1+4*i+1)+k], v[d*(1+4*i+3)+k], ctx), w, ctx);
-            t[4*i+0] = nmod_sub(t1, t4, ctx);
-            t[4*i+1] = nmod_sub(t2, t3, ctx);
-            t[4*i+2] = nmod_add(t1, t4, ctx);
-            t[4*i+3] = nmod_add(t2, t3, ctx);
-        }
-
-        Qrow = Q_;
-        for (i = 0; i < D; i++)
-        {
-            mp_limb_t c1h, c1m, c1;
-            mp_limb_t c2h, c2m, c2;
-            mp_limb_t c3h, c3m, c3;
-            mp_limb_t c4h, c4m, c4;
-            c1h = c1m = c1 = 0;
-            c2h = c2m = c2 = 0;
-            c3h = c3m = c3 = 0;
-            c4h = c4m = c4 = 0;
-            umul_ppmm(c4m, c4, Qrow[0], v[d*0+k]);
-            for (j = 0; j < D; j++)
-            {
-                MAC(c1h, c1m, c1, t[4*j+0], Qrow[4*j+1]);
-                MAC(c2h, c2m, c2, t[4*j+1], Qrow[4*j+2]);
-                MAC(c3h, c3m, c3, t[4*j+2], Qrow[4*j+3]);
-                MAC(c4h, c4m, c4, t[4*j+3], Qrow[4*j+4]);
-            }
-
-            Qrow += 4*D + 1;
-
-            NMOD_RED3(a[d*(4*i+1)+k], c1h, c1m, c1, ctx);
-            NMOD_RED3(a[d*(4*i+2)+k], c2h, c2m, c2, ctx);
-            NMOD_RED3(a[d*(4*i+3)+k], c3h, c3m, c3, ctx);
-            NMOD_RED3(a[d*(4*i+4)+k], c4h, c4m, c4, ctx);
-        }
-    }
-}
-
-
-void nmod_eval_interp_to_coeffs_n_fq(
-    mp_limb_t * a,
-    const mp_limb_t * v,
-    nmod_eval_interp_t E,
-    slong d,
-    nmod_t ctx)
-{
-    if (E->radix == 4)
-        _to_coeffs4_n_fq(a, v, E->T, E->Q, E->d, E->w, d, ctx);
-    else
-        _to_coeffs2_n_fq(a, v, E->T, E->Q, E->d, d, ctx);
-}
-
-void nmod_eval_interp_from_coeffs_n_fq(
-    mp_limb_t * v,
-    const mp_limb_t * a,
-    slong alen,
-    nmod_eval_interp_t E,
-    slong d,
-    nmod_t ctx)
-{
-    if (E->radix == 4)
-        _from_coeffs4_n_fq(v, a, alen, E->M, E->d, E->w, d, ctx);
-    else
-        _from_coeffs2_n_fq(v, a, alen, E->M, E->d, d, ctx);
-}
-
-
-void n_fq_evals_zero(n_fq_poly_t a)
-{
-    a->length = 0;
-}
-
-/* a += b */
-void n_fq_evals_add_inplace(
-    n_fq_poly_t a,
-    n_fq_poly_t b,
-    slong len,
-    const fq_nmod_ctx_t ctx)
-{
-    slong d = fq_nmod_ctx_degree(ctx);
-
-    if (b->length == 0)
-        return;
-
-    n_poly_fit_length(a, d*len);
-
-    if (a->length == 0)
-    {
-        _nmod_vec_set(a->coeffs, b->coeffs, d*len);
-        a->length = len;
-        return;
-    }
-
-    _nmod_vec_add(a->coeffs, a->coeffs, b->coeffs, d*len, ctx->mod);
-
-    a->length = _nmod_vec_is_zero(a->coeffs, d*len) ? 0 : len;
-}
-
-/* a = b*c */
-void n_fq_evals_mul(
-    n_fq_poly_t a,
-    n_fq_poly_t b,
-    n_fq_poly_t c,
-    slong len,
-    const fq_nmod_ctx_t ctx)
-{
-    slong d = fq_nmod_ctx_degree(ctx);
-    slong i;
-    mp_limb_t * tmp;
-    TMP_INIT;
-
-    if (b->length == 0 || c->length == 0)
-    {
-        a->length = 0;
-        return;
-    }
-
-    n_poly_fit_length(a, d*len);
-
-    TMP_START;
-
-    tmp = (mp_limb_t *) TMP_ALLOC(d*N_FQ_MUL_ITCH*sizeof(mp_limb_t));
-
-    for (i = 0; i < len; i++)
-        _n_fq_mul(a->coeffs + d*i, b->coeffs + d*i, c->coeffs + d*i, ctx, tmp);
-
-    a->length = _nmod_vec_is_zero(a->coeffs, d*len) ? 0 : len;
-
-    TMP_END;
-}
-
-/* a += b*c */
-void n_fq_evals_addmul(
-    n_fq_poly_t a,
-    n_fq_poly_t b,
-    n_fq_poly_t c,
-    slong len,
-    const fq_nmod_ctx_t ctx)
-{
-    slong d = fq_nmod_ctx_degree(ctx);
-    slong i;
-    mp_limb_t * tmp;
-    TMP_INIT;
-
-    if (b->length == 0 || c->length == 0)
-        return;
-
-    if (a->length == 0)
-    {
-        n_fq_evals_mul(a, b, c, len, ctx);
-        return;
-    }
-
-    TMP_START;
-
-    tmp = (mp_limb_t *) TMP_ALLOC(d*N_FQ_MUL_ITCH*sizeof(mp_limb_t));
-
-    for (i = 0; i < len; i++)
-        _n_fq_addmul(a->coeffs + d*i, a->coeffs + d*i,
-                            b->coeffs + d*i, c->coeffs + d*i, ctx, tmp);
-
-    a->length = _nmod_vec_is_zero(a->coeffs, d*len) ? 0 : len;
-
-    TMP_END;
-}
-
-void n_fq_evals_fmma(
-    n_fq_poly_t a,
-    n_fq_poly_t b,
-    n_fq_poly_t c,
-    n_fq_poly_t f,
-    n_fq_poly_t e,
-    slong len,
-    const fq_nmod_ctx_t ctx)
-{
-    slong d = fq_nmod_ctx_degree(ctx);
-    slong i;
-    mp_limb_t * tmp, * t;
-    TMP_INIT;
-
-    if (b->length == 0 || c->length == 0)
-    {
-        n_fq_evals_mul(a, f, e, len, ctx);
-        return;
-    }
-
-    if (f->length == 0 || e->length == 0)
-    {
-        n_fq_evals_mul(a, b, c, len, ctx);
-        return;
-    }
-
-    n_poly_fit_length(a, d*len);
-
-    TMP_START;
-
-    tmp = (mp_limb_t *) TMP_ALLOC(d*(1 + N_FQ_MUL_ITCH)*sizeof(mp_limb_t));
-    t = tmp + d*N_FQ_MUL_ITCH;
-
-    for (i = 0; i < len; i++)
-    {
-        _n_fq_mul(t, b->coeffs + d*i, c->coeffs + d*i, ctx, tmp);
-        _n_fq_addmul(a->coeffs + d*i, t, f->coeffs + d*i,
-                                          e->coeffs + d*i, ctx, tmp);
-    }
-
-    a->length = _nmod_vec_is_zero(a->coeffs, d*len) ? 0 : len;
-
-    TMP_END;
-}
-
-
-void nmod_eval_interp_to_coeffs_n_fq_poly(
-    n_fq_poly_t a,
-    const n_fq_poly_t v,
-    nmod_eval_interp_t E,
-    const fq_nmod_ctx_t ctx)
-{
-    slong d = fq_nmod_ctx_degree(ctx);
-    slong l = nmod_eval_interp_eval_length(E);
-
-    if (v->length == 0)
-    {
-        a->length = 0;
-        return;
-    }
-
-    FLINT_ASSERT(v->length == l);
-    n_poly_fit_length(a, d*l);
-    nmod_eval_interp_to_coeffs_n_fq(a->coeffs, v->coeffs, E, d, ctx->mod);
-    a->length = l;
-    _n_fq_poly_normalise(a, d);
-}
-
-void nmod_eval_interp_from_coeffs_n_fq_poly(
-    n_fq_poly_t v,
-    const n_fq_poly_t a,
-    nmod_eval_interp_t E,
-    const fq_nmod_ctx_t ctx)
-{
-    slong d = fq_nmod_ctx_degree(ctx);
-    slong l = nmod_eval_interp_eval_length(E);
-
-    if (a->length == 0)
-    {
-        v->length = 0;
-        return;
-    }
-
-    n_poly_fit_length(v, d*l);
-    v->length = l;
-    nmod_eval_interp_from_coeffs_n_fq(v->coeffs, a->coeffs, a->length, E, d, ctx->mod);
-}
-
 int n_fq_bpoly_hlift2_cubic(
     n_fq_bpoly_t A, /* clobbered (shifted by alpha) */
     n_fq_bpoly_t B0,
