+++ mpoly/monomials_deflation.c
@@ -9,7 +9,7 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
-#include "fmpz_mpoly.h"
+#include "mpoly.h"
 
 
 void mpoly_monomials_deflation(fmpz * shift, fmpz * stride,
