+++ nmod_mpoly_factor/n_bpoly_mod_factor_lgprime.c
@@ -13,7 +13,7 @@
 #include "fq_nmod_mpoly_factor.h"
 
 
-void n_bpoly_eval_fq_nmod_poly(
+static void n_bpoly_eval_fq_nmod_poly(
     fq_nmod_poly_t A,
     const fq_nmod_ctx_t ectx,
     const n_bpoly_t B)
@@ -38,7 +38,7 @@ void n_bpoly_eval_fq_nmod_poly(
 }
 
 
-void n_bpoly_mod_make_monic_mod(n_bpoly_t A, n_poly_t mk, nmod_t mod)
+static void n_bpoly_mod_make_monic_mod(n_bpoly_t A, n_poly_t mk, nmod_t mod)
 {
     slong i;
     n_poly_t t, lcinv;
@@ -64,7 +64,7 @@ void n_bpoly_mod_make_monic_mod(n_bpoly_t A, n_poly_t mk, nmod_t mod)
 }
 
 
-void n_bpoly_set_fq_nmod_poly_gen0(
+static void n_bpoly_set_fq_nmod_poly_gen0(
     n_bpoly_t A,
     const fq_nmod_poly_t B,
     const fq_nmod_ctx_t ectx)
@@ -79,7 +79,7 @@ void n_bpoly_set_fq_nmod_poly_gen0(
 }
 
 
-void n_bpoly_mod_mul_mod_poly(
+static void n_bpoly_mod_mul_mod_poly(
     n_bpoly_t A,
     const n_bpoly_t B,
     const n_bpoly_t C,
@@ -113,7 +113,7 @@ void n_bpoly_mod_mul_mod_poly(
 }
 
 /* division in ((Z/nZ)[y]/m(y))[x] */
-void n_bpoly_mod_divrem_mod_poly(
+static void n_bpoly_mod_divrem_mod_poly(
     n_bpoly_t Q,
     n_bpoly_t R,
     const n_bpoly_t A,
