+++ nmod_mpoly_factor/n_bpoly_mod_hlift.c
@@ -13,123 +13,6 @@
 
 #define FLINT_TMP_ARRAY_ALLOC(n, T) (T *) TMP_ALLOC(n*sizeof(T))
 
-void nmod_evals_zero(n_poly_t a)
-{
-    a->length = 0;
-}
-
-/* a += b */
-void nmod_evals_add_inplace(
-    n_poly_t a,
-    n_poly_t b,
-    slong len,
-    nmod_t ctx)
-{
-    slong i;
-
-    if (b->length == 0)
-        return;
-
-    n_poly_fit_length(a, len);
-
-    if (a->length == 0)
-    {
-        _nmod_vec_set(a->coeffs, b->coeffs, len);
-        a->length = len;
-        return;
-    }
-
-    for (i = 0; i < len; i++)
-        a->coeffs[i] = nmod_add(a->coeffs[i], b->coeffs[i], ctx);
-
-    a->length = _nmod_vec_is_zero(a->coeffs, len) ? 0 : len;
-}
-
-
-/* a = b*c */
-void nmod_evals_mul(
-    n_poly_t a,
-    n_poly_t b,
-    n_poly_t c,
-    slong len,
-    nmod_t ctx)
-{
-    slong i;
-
-    if (b->length == 0 || c->length == 0)
-    {
-        a->length = 0;
-        return;
-    }
-
-    n_poly_fit_length(a, len);
-
-    for (i = 0; i < len; i++)
-        a->coeffs[i] = nmod_mul(b->coeffs[i], c->coeffs[i], ctx);
-
-    a->length = _nmod_vec_is_zero(a->coeffs, len) ? 0 : len;
-}
-
-/* a += b*c */
-void nmod_evals_addmul(
-    n_poly_t a,
-    n_poly_t b,
-    n_poly_t c,
-    slong len,
-    nmod_t ctx)
-{
-    slong i;
-
-    if (b->length == 0 || c->length == 0)
-        return;
-
-    if (a->length == 0)
-    {
-        nmod_evals_mul(a, b, c, len, ctx);
-        return;
-    }
-
-    for (i = 0; i < len; i++)
-        NMOD_ADDMUL(a->coeffs[i], b->coeffs[i], c->coeffs[i], ctx);
-
-    a->length = _nmod_vec_is_zero(a->coeffs, len) ? 0 : len;
-}
-
-void nmod_evals_fmma(
-    n_poly_t a,
-    n_poly_t b,
-    n_poly_t c,
-    n_poly_t d,
-    n_poly_t e,
-    slong len,
-    nmod_t ctx)
-{
-    slong i;
-
-    if (b->length == 0 || c->length == 0)
-    {
-        nmod_evals_mul(a, d, e, len, ctx);
-        return;
-    }
-
-    if (d->length == 0 || e->length == 0)
-    {
-        nmod_evals_mul(a, b, c, len, ctx);
-        return;
-    }
-
-    n_poly_fit_length(a, len);
-
-    for (i = 0; i < len; i++)
-    {
-        mp_limb_t t = nmod_mul(b->coeffs[i], c->coeffs[i], ctx);
-        NMOD_ADDMUL(t, d->coeffs[i], e->coeffs[i], ctx);
-        a->coeffs[i] = t;
-    }
-
-    a->length = _nmod_vec_is_zero(a->coeffs, len) ? 0 : len;
-}
-
 
 int n_bpoly_mod_hlift2_cubic(
     n_bpoly_t A, /* clobbered (shifted by alpha) */
